- 회원 삭제, 탈퇴 기능 수정   - 동아리 회장일시 위임 후 탈퇴 가능 - 회원 관리 기능 수정   - 다른 사람을 회장으로 변경할 때 기존 회장의 등급을 일반회원으로 변경   - 삭제하려는 회원이 회장일시 삭제 불가능

+++ src/main/java/com/ac/su/clubmember/ClubMemberRestController.java
@@ -51,6 +51,10 @@ public class ClubMemberRestController {
         }
 
         // 회원 상태 수정
+        // 회원 상태를 회장으로 변경할시 기존에 있던 회장이 일반회원으로 변경됨
+        if(changeStatus == MemberStatus.CLUB_PRESIDENT) {
+            clubMemberService.delegatePresident(clubId);
+        }
         clubMemberService.changeStatus(memberId, clubId, changeStatus);
         return ResponseEntity.ok("회원 등급이 성공적으로 변경되었습니다.");
     }
@@ -71,7 +75,32 @@ public class ClubMemberRestController {
         }
 
         // 회원 삭제
-        clubMemberService.deleteMember(memberId, clubId);
-        return ResponseEntity.ok("회원을 탈퇴시켰습니다.");
+        // 삭제하려는 회원이 회장일시 삭제 불가능
+        MemberStatus targetStatus = clubMemberService.getMemberStatus(new ClubMemberId(memberId, clubId));
+        if (targetStatus == MemberStatus.CLUB_PRESIDENT) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ResponseMessage("동아리 회장 위임 후 탈퇴 가능합니다."));
+        } else {
+            clubMemberService.deleteMember(memberId, clubId);
+            return ResponseEntity.ok("회원을 탈퇴시켰습니다.");
+        }
+    }
+
+    // 회원이 자기 의지를 가지고 탈퇴
+    @PreAuthorize("isAuthenticated()")
+    @PostMapping("/{clubId}/withdrawClub")
+    public ResponseEntity<?> deleteMember(@PathVariable("clubId") Long clubId,
+                                          Authentication auth) {
+        // 회원 상태 가져오기
+        CustonUser user = (CustonUser) auth.getPrincipal();
+        MemberStatus status = clubMemberService.getMemberStatus(new ClubMemberId(user.getId(), clubId));
+
+        // 동아리 회장일시 위임 후 탈퇴 가능
+        if (status == MemberStatus.CLUB_PRESIDENT) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ResponseMessage("동아리 회장 위임 후 탈퇴 가능합니다."));
+        }
+
+        // 회원 탈퇴
+        clubMemberService.deleteMember(user.getId(), clubId);
+        return ResponseEntity.ok("동아리를 성공적으로 탈퇴했습니다.");
     }
 }
